Fix DOT (passing aDOT) integer-division truncation and show 2 decimals
The pass_depth_per_att_from_plays column computed SUM(dot) / SUM(attempts)
where both operands are integers, so Postgres did integer division and
truncated the result to a whole number before ROUND(...,2) ran (a no-op).
aDOT always rendered as e.g. 7 instead of 7.69. Cast the numerator to
numeric so the division keeps fractional precision, and bump the column's
display precision from 1 to 2 decimal places.

+++ libs-server/data-views-column-definitions/player-stats-from-plays-column-definitions.mjs
@@ -501,7 +501,7 @@ export default {
   }),
   player_pass_depth_per_pass_attempt_from_plays: player_stat_from_plays({
     pid_columns: ['psr_pid'],
-    with_select_string: `CASE WHEN SUM(CASE WHEN psr_pid IS NOT NULL AND (sk IS NULL OR sk = false) THEN 1 ELSE 0 END) > 0 THEN CAST(ROUND(SUM(dot) / SUM(CASE WHEN psr_pid IS NOT NULL AND (sk IS NULL OR sk = false) THEN 1 ELSE 0 END), 2) AS decimal) ELSE 0 END`,
+    with_select_string: `CASE WHEN SUM(CASE WHEN psr_pid IS NOT NULL AND (sk IS NULL OR sk = false) THEN 1 ELSE 0 END) > 0 THEN CAST(ROUND(SUM(dot)::numeric / SUM(CASE WHEN psr_pid IS NOT NULL AND (sk IS NULL OR sk = false) THEN 1 ELSE 0 END), 2) AS decimal) ELSE 0 END`,
     stat_name: 'pass_depth_per_att_from_plays',
     numerator_select: `SUM(dot)`,
     denominator_select: `SUM(CASE WHEN psr_pid IS NOT NULL AND (sk IS NULL OR sk = false) THEN 1 ELSE 0 END)`,
